add literal color value
